Add default UpdateCommand to use when updating model.
The UpdateCommand as well as the StoreCommand inspects the `$fillable`
attribute on the model to know which fields can be updated in the
resource.

It only takes non null fields but accepts empty fields, this allows
partial update to work, and it will not have problems to update a field
to an empty value.